@@ -78,100 +78,102 @@
     "nprogress": "^0.2.0",
     "path": "^0.12.7",
     "pinia": "^2.1.7",
-    "pinyin-pro": "^3.18.1",
+    "pinyin-pro": "^3.18.5",
     "qrcode": "^1.5.3",
     "qs": "^6.11.2",
     "responsive-storage": "^2.2.0",
-    "sortablejs": "^1.15.0",
-    "swiper": "^11.0.4",
-    "typeit": "^8.7.1",
+    "sortablejs": "^1.15.1",
+    "swiper": "^11.0.5",
+    "typeit": "^8.8.0",
     "v-contextmenu": "3.0.0",
     "v3-infinite-loading": "^1.3.1",
     "version-rocket": "^1.7.1",
-    "vue": "^3.3.8",
-    "vue-i18n": "^9.7.0",
-    "vue-json-pretty": "^2.2.4",
+    "vue": "^3.4.3",
+    "vue-i18n": "^9.8.0",
+    "vue-json-pretty": "^2.3.0",
     "vue-pdf-embed": "^1.2.1",
     "vue-router": "^4.2.5",
-    "vue-tippy": "^6.3.1",
+    "vue-tippy": "^6.4.1",
     "vue-types": "^5.1.1",
     "vue-virtual-scroller": "2.0.0-beta.8",
-    "vue-waterfall-plugin-next": "^2.2.8",
+    "vue-waterfall-plugin-next": "^2.3.1",
     "vue3-danmaku": "^1.6.0",
     "vuedraggable": "^4.1.0",
-    "wavesurfer.js": "^7.4.5",
-    "xgplayer": "^3.0.10",
+    "wavesurfer.js": "^7.6.0",
+    "xgplayer": "^3.0.11",
     "xlsx": "^0.18.5"
   },
   "devDependencies": {
-    "@commitlint/cli": "^18.4.2",
-    "@commitlint/config-conventional": "^18.4.2",
-    "@commitlint/types": "^18.4.0",
-    "@eslint/js": "^8.53.0",
+    "@commitlint/cli": "^18.4.3",
+    "@commitlint/config-conventional": "^18.4.3",
+    "@commitlint/types": "^18.4.3",
+    "@eslint/js": "^8.56.0",
     "@faker-js/faker": "^8.3.1",
     "@iconify-icons/ep": "^1.2.12",
     "@iconify-icons/ri": "^1.2.10",
     "@iconify/vue": "^4.1.1",
-    "@intlify/unplugin-vue-i18n": "^1.5.0",
+    "@intlify/unplugin-vue-i18n": "^2.0.0",
     "@pureadmin/theme": "^3.2.0",
-    "@types/intro.js": "^5.1.4",
+    "@types/intro.js": "^5.1.5",
     "@types/js-cookie": "^3.0.6",
-    "@types/node": "^20.9.0",
-    "@types/nprogress": "0.2.0",
+    "@types/node": "^20.10.6",
+    "@types/nprogress": "^0.2.3",
     "@types/qrcode": "^1.5.5",
-    "@types/qs": "^6.9.10",
-    "@types/sortablejs": "^1.15.5",
-    "@typescript-eslint/eslint-plugin": "^6.11.0",
-    "@typescript-eslint/parser": "^6.11.0",
-    "@vitejs/plugin-vue": "^4.5.0",
+    "@types/qs": "^6.9.11",
+    "@types/sortablejs": "^1.15.7",
+    "@typescript-eslint/eslint-plugin": "^6.16.0",
+    "@typescript-eslint/parser": "^6.16.0",
+    "@vitejs/plugin-vue": "^5.0.2",
     "@vitejs/plugin-vue-jsx": "^3.1.0",
     "autoprefixer": "^10.4.16",
     "cloc": "^2.11.0",
-    "cssnano": "^6.0.1",
-    "eslint": "^8.53.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-define-config": "^1.24.1",
-    "eslint-plugin-prettier": "^5.0.1",
-    "eslint-plugin-vue": "^9.18.1",
+    "cssnano": "^6.0.2",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-define-config": "^2.1.0",
+    "eslint-plugin-prettier": "^5.1.2",
+    "eslint-plugin-vue": "^9.19.2",
     "husky": "^8.0.3",
-    "lint-staged": "^15.1.0",
+    "lint-staged": "^15.2.0",
     "picocolors": "^1.0.0",
-    "postcss": "^8.4.31",
+    "postcss": "^8.4.32",
     "postcss-html": "^1.5.0",
     "postcss-import": "^15.1.0",
     "postcss-scss": "^4.0.9",
-    "prettier": "^3.1.0",
+    "prettier": "^3.1.1",
     "rimraf": "^5.0.5",
-    "rollup-plugin-visualizer": "^5.9.2",
-    "sass": "^1.69.5",
-    "stylelint": "^15.11.0",
-    "stylelint-config-recess-order": "^4.3.0",
+    "rollup-plugin-visualizer": "^5.12.0",
+    "sass": "^1.69.6",
+    "stylelint": "^16.1.0",
+    "stylelint-config-recess-order": "^4.4.0",
     "stylelint-config-recommended-vue": "^1.5.0",
-    "stylelint-config-standard-scss": "^11.1.0",
-    "stylelint-prettier": "^4.0.2",
-    "svgo": "^3.0.3",
-    "tailwindcss": "^3.3.5",
-    "typescript": "^5.2.2",
-    "vite": "^5.0.0",
+    "stylelint-config-standard-scss": "^12.0.0",
+    "stylelint-prettier": "^5.0.0",
+    "svgo": "^3.1.0",
+    "tailwindcss": "^3.4.0",
+    "typescript": "^5.3.3",
+    "vite": "^5.0.10",
     "vite-plugin-cdn-import": "^0.3.5",
     "vite-plugin-compression": "^0.5.1",
     "vite-plugin-fake-server": "^2.0.0",
-    "vite-plugin-remove-console": "^2.1.1",
+    "vite-plugin-remove-console": "^2.2.0",
     "vite-plugin-router-warn": "^1.0.0",
-    "vite-svg-loader": "^4.0.0",
+    "vite-svg-loader": "^5.1.0",
     "vue-eslint-parser": "^9.3.2",
-    "vue-tsc": "^1.8.22"
+    "vue-tsc": "^1.8.27"
   },
   "engines": {
-    "node": "^18.0.0 || >=20.0.0",
+    "node": "^18.12.0 || >=20.0.0",
     "pnpm": ">=8.6.10"
   },
   "packageManager": "pnpm@8.6.10",
   "pnpm": {
     "allowedDeprecatedVersions": {
       "sourcemap-codec": "*",
+      "domexception": "*",
       "w3c-hr-time": "*",
-      "stable": "*"
+      "stable": "*",
+      "abab": "*"
     }
   }
 }
