@@ -13,7 +13,8 @@ export interface iconType {
   align?: string;
   onLoad?: Function;
   includes?: Function;
-
-  //  all icon
+  // svg 需要什么SVG属性自行添加
+  fill?: string;
+  // all icon
   style?: object;
 }
